migrate generation logic from sbom to mod package
