+++ internal/cli/cmd/mod/testdata/integration/local.tar.gz
