+++ internal/cli/cmd/mod/testdata/integration/nested.tar.gz
