+++ internal/cli/cmd/mod/testdata/integration/no-dependencies.tar.gz
