+++ internal/cli/cmd/mod/testdata/integration/simple.tar.gz
