+++ internal/cli/cmd/mod/testdata/integration/snapshots/TestIntegrationLocal
