+++ internal/cli/cmd/mod/testdata/integration/vendored.tar.gz
